pyproject.toml
[tool.poetry]
name = "modlocale"
version = "1.0.0"
description = "ModLocale - 游戏模组本地化工具，用于提取和管理本地化字符串"
authors = ["Author Name <author@example.com>"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.0"
PyYAML = "^6.0.3"
tree-sitter = "^0.25.2"
tree-sitter-java = "^0.23.5"
tree-sitter-kotlin = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^9.0.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
flake8 = "^7.3.0"
pylint = "^3.3.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"
sphinx = "^7.4.0"
sphinx-rtd-theme = "^2.0.0"
sphinx-autodoc-typehints = "^2.3.0"
pre-commit = "^3.8.0"
pip-tools = "^7.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
exclude = ["__pycache__", ".venv", "build", "dist"]

[tool.pytest.ini_options]
python_files = "test_*.py"
testpaths = ["test"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true

[tool.black]
line-length = 120
include = "\\.pyi?$"
exclude = '''
/(\.eggs|\.git|\.hg|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist|__pycache__)/
'''

[tool.isort]
profile = "black"
line_length = 120
exclude = [".venv", "build", "dist", "__pycache__"]
